#8 Restructured to multimodule project for base API + extensions
