+++ holiday-calendar-base/src/main/java/com/github/davejoyce/calendar/FixedHoliday.java
