+++ holiday-calendar-base/src/main/lombok/com/github/davejoyce/calendar/FloatingHoliday.java
